Prepare 0.4.1 Maven Central release

+++ x-twitter-scraper-kotlin-core/src/main/kotlin/com/x_twitter_scraper/api/models/PaginatedUsers.kt
@@ -239,7 +239,7 @@ private constructor(
     internal fun validity(): Int =
         (if (hasNextPage.asKnown() == null) 0 else 1) +
             (if (nextCursor.asKnown() == null) 0 else 1) +
-            (users.asKnown()?.sumOf { it.validity().toInt() } ?: 0)
+            (users.asKnown()?.sumOf { it.validity() } ?: 0)
 
     override fun equals(other: Any?): Boolean {
         if (this === other) {
