+++ x-twitter-scraper-kotlin-core/src/main/kotlin/com/x_twitter_scraper/api/models/draws/DrawRetrieveResponse.kt
@@ -28,7 +28,9 @@ private constructor(
     @JsonCreator
     private constructor(
         @JsonProperty("draw") @ExcludeMissing draw: JsonField<DrawDetail> = JsonMissing.of(),
-        @JsonProperty("winners") @ExcludeMissing winners: JsonField<List<Winner>> = JsonMissing.of(),
+        @JsonProperty("winners")
+        @ExcludeMissing
+        winners: JsonField<List<Winner>> = JsonMissing.of(),
     ) : this(draw, winners, mutableMapOf())
 
     /**
@@ -200,8 +202,7 @@ private constructor(
      * Used for best match union deserialization.
      */
     internal fun validity(): Int =
-        (draw.asKnown()?.validity() ?: 0) +
-            (winners.asKnown()?.sumOf { it.validity().toInt() } ?: 0)
+        (draw.asKnown()?.validity() ?: 0) + (winners.asKnown()?.sumOf { it.validity() } ?: 0)
 
     override fun equals(other: Any?): Boolean {
         if (this === other) {
