+++ x-twitter-scraper-kotlin-core/src/main/kotlin/com/x_twitter_scraper/api/models/extractions/ExtractionRetrieveResponse.kt
@@ -34,7 +34,9 @@ private constructor(
         @JsonProperty("results")
         @ExcludeMissing
         results: JsonField<List<Result>> = JsonMissing.of(),
-        @JsonProperty("nextCursor") @ExcludeMissing nextCursor: JsonField<String> = JsonMissing.of(),
+        @JsonProperty("nextCursor")
+        @ExcludeMissing
+        nextCursor: JsonField<String> = JsonMissing.of(),
     ) : this(hasMore, job, results, nextCursor, mutableMapOf())
 
     /**
@@ -44,7 +46,7 @@ private constructor(
     fun hasMore(): Boolean = hasMore.getRequired("hasMore")
 
     /**
-     * Extraction job metadata — shape varies by tool type (JSON)
+     * Extraction job metadata - shape varies by tool type (JSON)
      *
      * @throws XTwitterScraperInvalidDataException if the JSON field has an unexpected type or is
      *   unexpectedly missing or null (e.g. if the server responded with an unexpected value).
@@ -145,7 +147,7 @@ private constructor(
          */
         fun hasMore(hasMore: JsonField<Boolean>) = apply { this.hasMore = hasMore }
 
-        /** Extraction job metadata — shape varies by tool type (JSON) */
+        /** Extraction job metadata - shape varies by tool type (JSON) */
         fun job(job: Job) = job(JsonField.of(job))
 
         /**
@@ -265,10 +267,10 @@ private constructor(
     internal fun validity(): Int =
         (if (hasMore.asKnown() == null) 0 else 1) +
             (job.asKnown()?.validity() ?: 0) +
-            (results.asKnown()?.sumOf { it.validity().toInt() } ?: 0) +
+            (results.asKnown()?.sumOf { it.validity() } ?: 0) +
             (if (nextCursor.asKnown() == null) 0 else 1)
 
-    /** Extraction job metadata — shape varies by tool type (JSON) */
+    /** Extraction job metadata - shape varies by tool type (JSON) */
     class Job
     @JsonCreator
     private constructor(
@@ -348,8 +350,9 @@ private constructor(
          *
          * Used for best match union deserialization.
          */
-        internal fun validity(): Int =
-            additionalProperties.count { (_, value) -> !value.isNull() && !value.isMissing() }
+        internal fun validity(): Int = additionalProperties.count { (_, value) ->
+            !value.isNull() && !value.isMissing()
+        }
 
         override fun equals(other: Any?): Boolean {
             if (this === other) {
@@ -445,8 +448,9 @@ private constructor(
          *
          * Used for best match union deserialization.
          */
-        internal fun validity(): Int =
-            additionalProperties.count { (_, value) -> !value.isNull() && !value.isMissing() }
+        internal fun validity(): Int = additionalProperties.count { (_, value) ->
+            !value.isNull() && !value.isMissing()
+        }
 
         override fun equals(other: Any?): Boolean {
             if (this === other) {
