+++ x-twitter-scraper-kotlin-core/src/main/kotlin/com/x_twitter_scraper/api/models/monitors/MonitorCreateParams.kt
@@ -295,7 +295,9 @@ private constructor(
             @JsonProperty("eventTypes")
             @ExcludeMissing
             eventTypes: JsonField<List<EventType>> = JsonMissing.of(),
-            @JsonProperty("username") @ExcludeMissing username: JsonField<String> = JsonMissing.of(),
+            @JsonProperty("username")
+            @ExcludeMissing
+            username: JsonField<String> = JsonMissing.of(),
         ) : this(eventTypes, username, mutableMapOf())
 
         /**
@@ -476,7 +478,7 @@ private constructor(
          * Used for best match union deserialization.
          */
         internal fun validity(): Int =
-            (eventTypes.asKnown()?.sumOf { it.validity().toInt() } ?: 0) +
+            (eventTypes.asKnown()?.sumOf { it.validity() } ?: 0) +
                 (if (username.asKnown() == null) 0 else 1)
 
         override fun equals(other: Any?): Boolean {
