+++ x-twitter-scraper-kotlin-core/src/main/kotlin/com/x_twitter_scraper/api/models/monitors/MonitorListResponse.kt
@@ -199,8 +199,7 @@ private constructor(
      * Used for best match union deserialization.
      */
     internal fun validity(): Int =
-        (monitors.asKnown()?.sumOf { it.validity().toInt() } ?: 0) +
-            (if (total.asKnown() == null) 0 else 1)
+        (monitors.asKnown()?.sumOf { it.validity() } ?: 0) + (if (total.asKnown() == null) 0 else 1)
 
     override fun equals(other: Any?): Boolean {
         if (this === other) {
