+++ x-twitter-scraper-kotlin-core/src/main/kotlin/com/x_twitter_scraper/api/models/support/tickets/TicketListResponse.kt
@@ -151,7 +151,7 @@ private constructor(
      *
      * Used for best match union deserialization.
      */
-    internal fun validity(): Int = (tickets.asKnown()?.sumOf { it.validity().toInt() } ?: 0)
+    internal fun validity(): Int = (tickets.asKnown()?.sumOf { it.validity() } ?: 0)
 
     class Ticket
     @JsonCreator(mode = JsonCreator.Mode.DISABLED)
