+++ x-twitter-scraper-kotlin-core/src/main/kotlin/com/x_twitter_scraper/api/models/webhooks/WebhookCreateParams.kt
@@ -476,7 +476,7 @@ private constructor(
          * Used for best match union deserialization.
          */
         internal fun validity(): Int =
-            (eventTypes.asKnown()?.sumOf { it.validity().toInt() } ?: 0) +
+            (eventTypes.asKnown()?.sumOf { it.validity() } ?: 0) +
                 (if (url.asKnown() == null) 0 else 1)
 
         override fun equals(other: Any?): Boolean {
