+++ x-twitter-scraper-kotlin-core/src/main/kotlin/com/x_twitter_scraper/api/models/webhooks/WebhookListResponse.kt
@@ -167,7 +167,7 @@ private constructor(
      *
      * Used for best match union deserialization.
      */
-    internal fun validity(): Int = (webhooks.asKnown()?.sumOf { it.validity().toInt() } ?: 0)
+    internal fun validity(): Int = (webhooks.asKnown()?.sumOf { it.validity() } ?: 0)
 
     override fun equals(other: Any?): Boolean {
         if (this === other) {
